temp-old/src/encontro10/monolito/requirements.txt
fastapi
uvicorn
rembg
pillow
python-multipart